--- flutter_news_app/ViewControllers/scratch.dart
@@ -1,1 +1,0 @@
-
