+++ flutter_news_app/lib/Base/HomeDataProvider.dart
@@ -1,5 +1,6 @@
 
 import 'package:flutter_news_app/Base/RoutParam.dart';
+import 'package:flutter_news_app/Section/Async/AsyncRoute.dart';
 import 'package:flutter_news_app/Section/Interface/ConstValue.dart';
 import 'package:flutter_news_app/Section/Merterial/Merterial.dart';
