--- flutter_news_app/lib/gen_a/A.dart
@@ -1,4 +1,0 @@
-class A{
- //auto gen ,do net edit !!! 
-  static String test="test";
-}
